// apps/portal/src/api/generated/model/propertyOverviewResponseDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */
import type { InvitesOverviewDto } from "./invitesOverviewDto";
import type { OwnersOverviewDto } from "./ownersOverviewDto";
import type { UnitsOverviewDto } from "./unitsOverviewDto";

export interface PropertyOverviewResponseDto {
    units: UnitsOverviewDto;
    owners: OwnersOverviewDto;
    invites: InvitesOverviewDto;
}